Added embedded screenshots - reverted back to opening/closing browser :(

+++ src/main/java/utils/DriverUtils.java
@@ -1,5 +1,7 @@
 package utils;
 
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 
@@ -8,22 +10,42 @@ import java.util.concurrent.TimeUnit;
 public class DriverUtils {
     private static WebDriver driver;
 
+    private static WebDriver startWebDriver() {
+        System.out.println("Start WebDriver");
+        WebDriver d = new FirefoxDriver();
+        d.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
+        d.manage().timeouts().pageLoadTimeout(60, TimeUnit.SECONDS);
+        d.manage().timeouts().setScriptTimeout(60,TimeUnit.SECONDS);
+        d.manage().window().maximize();
+        return d;
+    }
+
+    public static byte[] takeScreenShot()
+    {
+        return ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
+    }
+
+    public static WebDriver resetWebDriver() {
+        if (driver == null) {
+            driver = startWebDriver();
+        }
+        else
+        {
+            driver.manage().deleteAllCookies();
+        }
+        return driver;
+    }
+
     public static WebDriver getDriver() {
         if (driver == null) {
-            System.out.println("Init WebDriver");
-            driver = new FirefoxDriver();
-            driver.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
-            driver.manage().timeouts().pageLoadTimeout(60, TimeUnit.SECONDS);
-            driver.manage().timeouts().setScriptTimeout(60,TimeUnit.SECONDS);
-            driver.manage().window().maximize();
+            driver = startWebDriver();
         }
-        driver.manage().deleteAllCookies();
         return driver;
     }
 
     public static void quitDriver() {
         System.out.println("Quit WebDriver");
-        driver.quit();
+        if (driver != null) driver.quit();
         driver = null;
     }
 }
